+++ src/main/java/org/logstash/dissect/DissectResult.java
@@ -1,0 +1,21 @@
+package org.logstash.dissect;
+
+public class DissectResult {
+    private boolean bailed;
+
+    public DissectResult() {
+        bailed = false;
+    }
+
+    public final void bail() {
+        bailed = true;
+    }
+
+    public boolean matched() {
+        return !bailed;
+    }
+
+    public boolean notMatched() {
+        return bailed;
+    }
+}