+++ src/main/java/org/logstash/dissect/fields/FieldComparator.java
@@ -6,6 +6,6 @@ import java.util.Comparator;
 public final class FieldComparator implements Comparator<Field>, Serializable {
     @Override
     public int compare(final Field o1, final Field o2) {
-        return o1.ordinal() - o2.ordinal();
+        return Integer.compare(o1.ordinal(), o2.ordinal());
     }
 }