+++ src/main/java/org/logstash/dissect/fields/FieldFactory.java
@@ -41,11 +41,14 @@ import org.logstash.dissect.Delimiter;
 public final class FieldFactory {
     private static final String MIXED_PREFIXES = "Field cannot prefix with both Append and Indirect Prefix (%s): %s";
     private static final String PREFIXED_EMPTY = "Field cannot be a prefix on its own without further text";
+    private static final String NAME_EMPTY = "Field cannot be a prefix and a suffix without a name section";
+
+    private FieldFactory() {
+    }
 
     public static Field create(final int id, final String field, final Delimiter previous, final Delimiter next) {
         final String[] nameSuffix;
-
-        if (field.isEmpty() || field.startsWith("?")) {
+        if (field.isEmpty() || field.charAt(0) == '?') {
             nameSuffix = Field.extractNameSuffix(removeLeadingCharIfPresent(field));
             return SkipField.create(id, nameSuffix[0], nameSuffix[1], previous, next);
         }
@@ -55,20 +58,27 @@ public final class FieldFactory {
         if (field.startsWith("&+")) {
             throw new InvalidFieldException(String.format(MIXED_PREFIXES, "&+", field));
         }
-        if (field.startsWith("+")) {
-            final String shorterField = removeLeadingCharIfPresent(field);
+        final String shorterField;
+        if (field.charAt(0) == '+') {
+            shorterField = removeLeadingCharIfPresent(field);
             if (shorterField.isEmpty()) {
                 throw new InvalidFieldException(PREFIXED_EMPTY);
             }
             nameSuffix = Field.extractNameSuffix(shorterField);
+            if (nameSuffix[0].isEmpty()) {
+                throw new InvalidFieldException(NAME_EMPTY);
+            }
             return AppendField.create(id, nameSuffix[0], nameSuffix[1], previous, next);
         }
-        if (field.startsWith("&")) {
-            final String shorterField = removeLeadingCharIfPresent(field);
+        if (field.charAt(0) == '&') {
+            shorterField = removeLeadingCharIfPresent(field);
             if (shorterField.isEmpty()) {
                 throw new InvalidFieldException(PREFIXED_EMPTY);
             }
             nameSuffix = Field.extractNameSuffix(shorterField);
+            if (nameSuffix[0].isEmpty()) {
+                throw new InvalidFieldException(NAME_EMPTY);
+            }
             return IndirectField.create(id, nameSuffix[0], nameSuffix[1], previous, next);
         }
         nameSuffix = Field.extractNameSuffix(field);