+++ src/main/java/org/logstash/dissect/search/MultiByteLocator.java
@@ -7,12 +7,9 @@ public class MultiByteLocator implements DelimiterLocator {
     }
 
     @Override
-    public int indexOf(final byte[] needle, final byte[] haystack, final int offset) {
+    public final int indexOf(final byte[] needle, final byte[] haystack, final int offset) {
         int localOffset = offset;
-
-        final int sourceOffset = 0;
         final int sourceCount = haystack.length;
-        final int targetOffset = 0;
 
         if (localOffset >= sourceCount) {
             return -1;
@@ -21,10 +18,10 @@ public class MultiByteLocator implements DelimiterLocator {
             localOffset = 0;
         }
 
-        final byte first = needle[targetOffset];
-        final int max = sourceOffset + (sourceCount - needle.length);
+        final byte first = needle[0];
+        final int max = sourceCount - needle.length;
 
-        for (int i = sourceOffset + localOffset; i <= max; i++) {
+        for (int i = localOffset; i <= max; i++) {
             /* Look for first byte. */
             if (haystack[i] != first) {
                 while (++i <= max && haystack[i] != first) {
@@ -36,14 +33,14 @@ public class MultiByteLocator implements DelimiterLocator {
             if (i <= max) {
                 int j = i + 1;
                 final int end = j + needle.length - 1;
-                for (int k = targetOffset + 1; j < end && haystack[j]
+                for (int k = 1; j < end && haystack[j]
                         == needle[k]; j++, k++) {
                     // do nothing
                 }
 
                 if (j == end) {
                     /* Found whole pattern. */
-                    return i - sourceOffset;
+                    return i;
                 }
             }
         }